task 16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,21 +8,21 @@
 
 int main(void)
 {
-	unsigned long int n1 = 1, n2 = 2, next;
-	int count;
+	unsigned long long  n1 = 1, n2 = 2, next;
+	unsigned long long  count;
 
-	printf("%lu, %lu, ", n1, n2);
+	printf("%llu, %llu, ", n1, n2);
 
 	for (count = 3; count <= 98; count++)
 	{
 		next = n1 + n2;
 		if (count < 98)
 		{
-			printf("%lu, ", next);
+			printf("%llu, ", next);
 		}
 		else
 		{
-			printf("%lu\n", next);
+			printf("%llu\n", next);
 		}
 		n1 = n2;
 		n2 = next;
